requirements.txt
annotated-types==0.7.0
anyio==4.10.0
apturl==0.5.2
asgiref==3.8.1
backports-datetime-fromisoformat==2.0.3
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.2.1
colorama==0.4.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.1
django-jazzmin==3.0.1
dnspython==2.8.0
duplicity==0.8.21
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
fasteners==0.14.1
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
future==0.18.2
greenlet==3.2.4
h11==0.16.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==6.0.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
monotonic==1.6
more-itertools==8.10.0
mysqlclient==2.2.7
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==25.0
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
pipenv==2025.0.2
platformdirs==4.3.8
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.6.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.15.3
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-jose==3.5.0
python-pptx==1.0.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.4.2
starlette==0.47.2
systemd-python==234
typing-inspection==0.4.1
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.35.0
virtualenv==20.31.2
wadllib==1.3.6
Werkzeug==3.1.3
WTForms==3.2.1
xdg==5
xkit==0.0.0
xlsxwriter==3.2.5
zipp==1.0.0
